Add stages

+++ basic/controllers/TourStageController.php
@@ -2,17 +2,24 @@
 
 namespace app\controllers;
 
+use app\helpers\LoggerTrait;
+use app\models\Tour;
 use app\models\TourStage;
 use app\models\TourStageSearch;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
+use yii\filters\AccessControl;
 use yii\filters\VerbFilter;
+use Yii;
 
 /**
  * TourStageController implements the CRUD actions for TourStage model.
  */
 class TourStageController extends Controller
 {
+    use LoggerTrait;
+
     /**
      * @inheritDoc
      */
@@ -21,6 +28,16 @@ class TourStageController extends Controller
         return array_merge(
             parent::behaviors(),
             [
+                'access' => [
+                    'class' => AccessControl::class,
+                    'only' => ['create', 'update', 'delete'],
+                    'rules' => [
+                        [
+                            'allow' => true,
+                            'roles' => ['@'],
+                        ],
+                    ],
+                ],
                 'verbs' => [
                     'class' => VerbFilter::className(),
                     'actions' => [
@@ -32,54 +49,43 @@ class TourStageController extends Controller
     }
 
     /**
-     * Lists all TourStage models.
-     *
-     * @return string
-     */
-    public function actionIndex()
-    {
-        $searchModel = new TourStageSearch();
-        $dataProvider = $searchModel->search($this->request->queryParams);
-
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
-    }
-
-    /**
-     * Displays a single TourStage model.
-     * @param int $id ID
-     * @return string
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionView($id)
-    {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
-    }
-
-    /**
-     * Creates a new TourStage model.
+     * Add a new TourStage model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return string|\yii\web\Response
      */
-    public function actionCreate()
+    public function actionAddStage($tour_id)
     {
-        $model = new TourStage();
-
-        if ($this->request->isPost) {
-            if ($model->load($this->request->post()) && $model->save()) {
-                return $this->redirect(['view', 'id' => $model->id]);
+        if(Yii::$app->user->can('tourCrud')){
+            $tour = Tour::findOne($tour_id);
+            if (!$tour) {
+                throw new NotFoundHttpException("Tour not found.");
             }
-        } else {
-            $model->loadDefaultValues();
-        }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+            $model = new TourStage();
+            $model->tour_id = $tour->id;
+
+            $nextSequence = TourStage::find()
+                ->where(['tour_id' => $tour_id])
+                ->max('sequence');
+            $model->sequence = $nextSequence ? $nextSequence + 1 : 1;
+
+            if ($this->request->isPost) {
+                if ($model->load($this->request->post()) && $model->save()) {
+                    Yii::$app->session->setFlash('success', 'Stage added successfully.');
+                    $this->logInfo('Created tour stage', ['model' => $model, 'user' => Yii::$app->user->identity->license]);
+                    return $this->redirect(['tour/view', 'id' => $tour_id]);
+                }
+            } else {
+                $model->loadDefaultValues();
+            }
+
+            return $this->render('add', [
+                'model' => $model,
+                'tour' => $tour,
+            ]);
+        } else {
+            throw new ForbiddenHttpException();
+        }
     }
 
     /**
@@ -91,15 +97,20 @@ class TourStageController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can('tourCrud')){
+            $model = $this->findModel($id);
 
-        if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
+                $this->logInfo('Updated tour stage', ['model' => $model, 'user' => Yii::$app->user->identity->license]);
+                return $this->redirect(['view', 'id' => $model->id]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException();
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -111,9 +122,18 @@ class TourStageController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+        if(Yii::$app->user->can('tourCrud')){
+            $model = $this->findModel($id);
+            if($model->flights) {
+                Yii::$app->session->setFlash('error', 'Can\'t delete stage with flights associated.');
+            } else {
+                $this->findModel($id)->delete();
+                $this->logInfo('Deleted tour stage', ['id' => $id, 'user' => Yii::$app->user->identity->license]);
+            }
+            return $this->redirect(['tour/view', 'id' => $model->tour->id]);
+        } else {
+            throw new ForbiddenHttpException();
+        }
     }
 
     /**
